Create TextBox widget
Created a TextBox widget for inputing text. The text is represented
internally as std::wstring for UTF-16 support. Characters within the
TextBox can be releted using the backspace and delete keys. The left and
right arrow keys can be used to navigate the contents.

A Listener class handles callback events. The TextBox stores a list of
Listeners that can be added and removed.